==> 1_DataCleaning_EDA.ipynb <==
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
from pyspark.sql.types import *
from pyspark.sql.functions import from_unixtime
from pyspark.ml.recommendation import ALS
from pyspark.ml.feature import StringIndexer, IndexToString
import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns
from scipy import stats
from pyspark.sql.functions import col, sum as _sum
from pyspark.sql.functions import col, regexp_replace, length

## Raw Data Extraction from HDFS with PySpark

from pyspark.sql import SparkSession

spark = SparkSession.builder \
    .appName("SteamReviewsHDFS") \
    .config("spark.hadoop.fs.defaultFS", "hdfs://localhost:9000") \
    .config("spark.driver.memory", "4g") \
    .config("spark.executor.memory", "4g") \
    .config("spark.sql.shuffle.partitions", "100")\
    .config("spark.sql.execution.arrow.pyspark.enabled", "true") \
    .getOrCreate()


df = spark.read.csv("/user/tejashree/project/data/raw/steam_reviews.csv", header=True, inferSchema=True)


df.show(5)


df.printSchema()

df.count()

## Data Cleaning and Pre-Processing

- Convert the column names into valid names - remove the dots and replace them with underscore.
- Remove the NULL records from the data set 
- Convert the columns into appropriate data types. For e.g., app_id can be converted to IntegerType() and all the timestamp strings into timestamps.

data_csv = df.withColumnRenamed("author.steamid","author_steamid"). \
withColumnRenamed("author.num_games_owned","author_num_games_owned").\
withColumnRenamed("author.num_reviews", "author_num_reviews").\
withColumnRenamed("author.playtime_forever","author_playtime_forever").\
withColumnRenamed("author.playtime_at_review","author_playtime_at_review").\
withColumnRenamed("author.last_played","author_last_played").\
withColumnRenamed("author.playtime_last_two_weeks", "author_playtime_last_two_weeks")

data_csv.printSchema()

## Check NULL values

columns_to_check = [c for c in data_csv.columns if c != '_c0']

null_counts = data_csv.select([
    _sum(col(c).isNull().cast("int")).alias(c) for c in columns_to_check
])

null_counts_pd = null_counts.toPandas()
print(null_counts_pd.T)  # T = transpose


## Handling missing values
1. Drop rows missing app_id, review_id, review, recommended (core fields)
2. Fill numeric columns like votes_helpful, votes_funny with 0
3. Ignore user info if missing (optional for dashboard)



important_cols = ["app_id","app_name", "review_id", "review", "recommended","author_steamid"]
numeric_fill_cols = ["votes_helpful", "votes_funny", "comment_count", "weighted_vote_score"]

#Drop rows where any important column is NULL
df_clean = data_csv.dropna(subset=important_cols)

#Fill numeric columns (votes helpful, funny, comment_count) with 0 where NULL
for col_name in numeric_fill_cols:
    df_clean = df_clean.withColumn(col_name, coalesce(col_name, lit(0)))

#fill steam_purchase, received_for_free to 'Unknown'
df_clean = df_clean.fillna({
    "steam_purchase": "unknown",
    "received_for_free": "unknown"
})

## Normalize text
1. Remove HTML tags
2. Remove emojis and non-text symbols
3. Keep Unicode letters (all languages) and spaces

#  Normalize text:
#    a. Remove HTML tags
#    b. Remove emojis and non-text symbols
#    c. Keep Unicode letters (all languages) and spaces
df_clean = df_clean.withColumn(
    "cleaned_review",
    regexp_replace(
        lower(col("review")),
        "<[^>]*>",  # Remove HTML
        ""
    )
)
df_clean = df_clean.withColumn(
    "cleaned_review",
    regexp_replace(
        col("cleaned_review"),
        "[^\\p{L}\\s]",  # Keep Unicode letters and whitespace only
        ""
    )
)
df_clean = df_clean.withColumn("cleaned_review", trim(col("cleaned_review")))


df_sample = df_clean.limit(10).toPandas()
print(df_sample)

df_clean.count()

df_clean.printSchema()

## Changing the data type of each columns into appropriate types

# Changing the data type of each columns into appropriate types
df_clean = df_clean.withColumn("app_id",df_clean["app_id"].cast(IntegerType())).\
            withColumn("review_id", df_clean["review_id"].cast(LongType())).\
            withColumn("author_steamid", df_clean["author_steamid"].cast(LongType())).\
            withColumn("recommended", df_clean["recommended"].cast(BooleanType())).\
            withColumn("steam_purchase", df_clean["steam_purchase"].cast(BooleanType())).\
            withColumn("author_num_games_owned", df_clean["author_num_games_owned"].cast(IntegerType())).\
            withColumn("author_num_reviews", df_clean["author_num_reviews"].cast(IntegerType())).\
            withColumn("author_playtime_forever", df_clean["author_playtime_forever"].cast(FloatType())).\
            withColumn("author_playtime_at_review", df_clean["author_playtime_at_review"].cast(FloatType()))

# Converting the time columns into timestamp data type
df_clean = df_clean.withColumn("timestamp_created", from_unixtime("timestamp_created").cast("timestamp")).\
            withColumn("author_last_played", from_unixtime(df_clean["author_last_played"]).cast(TimestampType())).\
            withColumn("timestamp_updated", from_unixtime(df_clean["timestamp_updated"]).cast(TimestampType()))

df_clean.printSchema()


# Clean app_name,review: Keep only rows where app_name is NOT purely numeric AND review is NOT purely numeric



# Filter: Keep only rows where app_name is NOT purely numeric AND review is NOT purely numeric
df_clean = df_clean.filter(
    (~col("app_name").rlike("^[0-9]+$")) &    # app_name is NOT all numbers
    (~col("review").rlike("^[0-9]+$")) &    # app_name is NOT all numbers
    (~col("cleaned_review").rlike("^[0-9]+$"))            # review is NOT all numbers
)

# Optional: Check the result
df_clean.select("app_name", "review","cleaned_review").show(10, truncate=False)


df_clean.count()

df_sample2 = df_clean.limit(10).toPandas()
print(df_sample2)

## Save the cleaned data to HDFS in Parquet format 

df_clean.write.mode("overwrite").parquet("/user/tejashree/project/data/processed/cleaned_steam_reviews.parquet")

# Try to read the parquet back
try:
    test_df = spark.read.parquet("/user/tejashree/project/data/processed/cleaned_steam_reviews.parquet")
    print("Parquet read successful. Number of rows:", test_df.count())
except Exception as e:
    print("Error reading parquet:", str(e))


## Filter data for Sentiment analysis langauga = english

df_cleaned_parquet = spark.read.parquet("/user/tejashree/project/data/processed/cleaned_steam_reviews.parquet")


english_df = df_cleaned_parquet.filter(df_cleaned_parquet.language == "english")


english_df.write.mode("overwrite").parquet("/user/tejashree/project/data/processed/steam_review_english.parquet")


try:
    test_df = spark.read.parquet("/user/tejashree/project/data/processed/steam_review_english.parquet")
    print("Parquet read successful. Number of rows:", test_df.count())
except Exception as e:
    print("Error reading parquet:", str(e))


spark.stop()
