==> 5_Game_Sentiment_summary.ipynb <==
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
from pyspark.sql.types import *
from pyspark.sql.functions import from_unixtime
from pyspark.ml.recommendation import ALS
from pyspark.ml.feature import StringIndexer, IndexToString
import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns
from scipy import stats
from pyspark.sql.functions import col, sum as _sum
from pyspark.sql import SparkSession
from pyspark.sql.functions import udf
from pyspark.sql.types import StringType

from pyspark.sql import SparkSession

spark = SparkSession.builder \
    .appName("SteamReviewsHDFStransalte") \
    .config("spark.hadoop.fs.defaultFS", "hdfs://localhost:9000") \
    .config("spark.driver.memory", "4g") \
    .config("spark.executor.memory", "4g") \
    .config("spark.sql.shuffle.partitions", "100")\
    .config("spark.sql.execution.arrow.pyspark.enabled", "true") \
    .getOrCreate()



df = spark.read.parquet("/user/tejashree/project/outputs/steam_sentiment_final_batched.parquet") 

# Add numeric sentiment for easier aggregation
df = df.withColumn("is_positive", when(col("sentiment_label") == "POSITIVE", 1).otherwise(0))
df = df.withColumn("is_negative", when(col("sentiment_label") == "NEGATIVE", 1).otherwise(0))


# Group by app_id and app_name
summary = df.groupBy("app_id", "app_name").agg(
    avg("sentiment_score").alias("avg_sentiment_score"),
    count("*").alias("total_reviews"),
    count(when(col("is_positive") == 1, True)).alias("positive_reviews"),
    count(when(col("is_negative") == 1, True)).alias("negative_reviews")
)

# Calculate percentages
summary = summary.withColumn("percent_positive", (col("positive_reviews") / col("total_reviews")) * 100)
summary = summary.withColumn("percent_negative", (col("negative_reviews") / col("total_reviews")) * 100)


summary.show(5, truncate=False)

summary.filter(col("app_name") == "Hades").show(5, truncate=False)


# Save summary table
summary.write.mode("overwrite").parquet("/user/tejashree/project/data/processed/steam_sentiment_summary.parquet")

print("✅ Sentiment summary saved to /user/tejashree/project/data/processed/steam_sentiment_summary.parquet")
